Update function documentation with array functions
There are now two functions one for an ESC on a specific Adafruit PCA9685 I2C 16-Channel PWM/Servo Driver pin, and one for the whole array of ESCs attached to the Adafruit PCA9685 I2C 16-Channel PWM/Servo Driver.

Constructor no longer uses pin inputs as users are expected to use the two pairs of pins in order for the controls and reverse signal lines.